fix: protect working directory from deleting

+++ R/qti_test.R
@@ -32,8 +32,10 @@ create_qti_test <- function(object, path = getwd(), verification = FALSE) {
 
     if (!dir.exists(dir)) dir.create(dir, recursive = TRUE)
     # to remove content of dir
+    if (path != getwd()) {
     file_list <- list.files(dir, full.names = TRUE)
     file.remove(file_list)
+    }
 
     content <- createAssessmentTest(object, dir)
     doc_test <- xml2::read_xml(as.character(content))
